+++ lifelines/utils/__init__.py
@@ -98,29 +98,6 @@ class CensoringType:
             return
 
 
-class StatError(Exception):
-    pass
-
-
-class ConvergenceError(ValueError):
-    # inherits from ValueError for backwards compatibility reasons
-    def __init__(self, msg, original_exception=""):
-        super(ConvergenceError, self).__init__(msg + "%s" % original_exception)
-        self.original_exception = original_exception
-
-
-class ConvergenceWarning(RuntimeWarning):
-    pass
-
-
-class StatisticalWarning(RuntimeWarning):
-    pass
-
-
-class ApproximationWarning(RuntimeWarning):
-    pass
-
-
 def qth_survival_times(q, survival_functions) -> Union[pd.DataFrame, float]:
     """
     Find the times when one or more survival functions reach the qth percentile.